Cleanup `ILoadoutViewModel`
